Use simpler package file tooling.
This change mimics the pip-compile workflow, but for R.

packages.in is used to generate a fully pinned dependency tree in packages.txt,
which is used to install R packages in the image.

The initial packages.in was hand compiled from the list of packages in the old
Dockerfile, pinned to their most recent release. A handful were dropped from
the explicit list, as they are already included in the r-essentials package,
which is the recommended base package for R in conda.

There is tooling to update packages.txt from packages.in, which in an
inception-like manner uses the docker image produced by this repo to do it...