Add dtwclust package
This required squashing the image to a single layer, as we'd hit the
limit of 128 layers. The squash.sh script included here performs this.

If we add another 126 packages and need to squash again before we
properly fix the R image, I will cry.

It *also* required installing from a repo with a specific date. This
required some modifications to the add-pkg.sh script to take an optional
repo parameter