fix: avoid creating renv directory in tests.
Whenever we run R in the container, it activates renv, which AFAICT will
*always* create an `renv` directory in the current.

The test.sh script was mounting this directory in, thus creating an renv
dir in it, so we stop doing that.

Additionally, add renv/ dir to the gitignore.

In the processs, realised this dir creation created a sandbox dir, which
could be instead done once in /renv, so made that change too.

Ideally, we'd tell renv to create the dir somewhere else, but there
doesn't seem to be away to do that, it's hardcoded to be `./renv`.

Also, apply suggested docs fix from Tom.