Strip dependencies
Figure out the actual dependencies needed, rather than the -dev or -bin
versions. This removes 130 uneeded packages, and saves 230mb of the
image.

The only testing done here is that the package imports cleanly, which
definitly does fail if you remove the dependency entirely. It's possible
that we have removed some optional dependencies that are not loaded on
package import, but are are needed for certain. However, it is easy to
add these back again if needed.