Remove pak_cleanup() call
AFAICT, this has no effect, for two reasons.

 - ENV vars are only valid in the build stage that defined them, and we
   defined them in the `base-r` stage. In the `r` stage it will run with
   the default values, which should be empty anyway.

 - You cannot save space by deleting things with docker. The objects
   will still be in the previous layer.  The only way to reduce space by
   deleting file by removing them within the same RUN command that
   created them, so they are never persisted to a layer.

In addition, because we use a persistant build cache, we don't want to
clear it at all! That would wipe the persistant cache out each build.

I manually checked that this made no different to the size, and it did
not, and I poked around to check:

```
> library("pak")
> cache_summary()
$cachepath
[1] "/root/.cache/R/pkgcache/pkg"

$files
[1] 0

$size
[1] 0

> cache_list()
> A data frame: 0 × 6
> ℹ 6 variables: fullpath <chr>, path <chr>, package <chr>, url <chr>,
>   etag <chr>, sha256 <chr>
```