Make v2
This adds version 2 of the r image in the v2 directory. The packages for
the image are specified in v2/packages.toml. The structure of this file
is explained in the README. The rationale for the approach taken in the
v2 image is explained in the README, as is how to add a new package and
update the image. At the end of the build process the v2/pkg.lock and
v2/packages.md files both record the included packages and their
versions, the latter of which is meant for convenient reading for users
on GitHub.