Make v1, create packages.md, and move and expand tests
This moves the machinery for the current r image into the v1 directory
in preparation for adding version 2 in a v2 directory.

This also adds the scripts/packages.Rmd R Markdown document, shell
script, and justfile recipe to create the v1/packages.md file so that
users can conveniently see the packages and their versions included in
the image.

This also moves the tests to the tests directory and adds additional
tests. The tests now report the number of installed R packages, test the
loading of the base R packages (in addition to all the added packages),
report the capabilities of the arrow package, and check that a call to
Rcpp::evalCpp() runs successfully (i.e., this checks that a relevant
compiler is available).